# docs/swagger.yaml
basePath: /
definitions:
...
  model.Address:
    properties:
      city:
        type: string
      country:
        type: string
      pincode:
        type: integer
      state:
        type: string
      street:
        type: string
    type: object
...
  model.Bookdetails:
    properties:
      book_id:
        type: string
      time_taken:
        type: integer
      timepenaltycalc:
        type: integer
      title:
        type: string
    required:
    - title
    type: object
...
  model.Logout:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  model.User:
    properties:
      _id:
        type: string
      address:
        $ref: '#/definitions/model.Address'
      books_taken:
        items:
          $ref: '#/definitions/model.Bookdetails'
        type: array
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      isfirstlogin:
        type: boolean
      last_name:
        type: string
      login:
        type: boolean
      mobile_no:
        type: string
      password:
        type: string
      status:
        type: string
      total_penalty:
        type: integer
      user_type:
        type: string
      username:
        type: string
    required:
    - email
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a  Library Management API server.
  title: Library Management API
  version: "1.0"
paths:
...
  /users/login:
    post:
      consumes:
      - application/json
      operationId: user-logout
      parameters:
      - description: Payload for logout API
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.Logout'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                type: string
              type: array
            type: object
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: logout user
  /users/signup:
    post:
      consumes:
      - application/json
      operationId: user-signup
      parameters:
      - description: Payload for Signup API
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.User'
        "400":
          description: Bad Request
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: signup user
swagger: "2.0"